♻️ refactor(backendv2): split export artifact writers
