+++ tests/backendv2/test_routing_v2_raw_weights.py
@@ -1,3 +1,4 @@
+# ruff: noqa: E402
 import numpy as np
 import paicorelib
 import torch
